move getheader to bar
